+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.8.2
 
+* Various Bugfixes
 * Migrated to SCSS
 * Migrated to AngularJS
 * Migrated to CoffeeScript
@@ -8,6 +9,7 @@
 * Integrated Grunt
 * Using Videogular as audio/video player
 * Episodes are now marked as played when reaching the last minute of playback
+* Added placeholder cover
 
 0.8.1